browser and embark entrypoints go through index
